Interface created

+++ Backend/service-ticket-tool/src/main/java/com/sts/repository/UserRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sts.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long>
+{
 
 }
